cliving/requirements.txt
absl-py==2.1.0
asgiref==3.8.0
attrs==23.2.0
boto3==1.34.127
botocore==1.34.127
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
dj-rest-auth==6.0.0
Django==5.0.3
django-allauth==64.2.1
django-cors-headers==4.4.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
ffmpeg-python==0.2.0
filelock==3.15.1
fire==0.6.0
flatbuffers==24.3.25
flex==6.14.1
fonttools==4.51.0
fsspec==2024.6.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.64.1
huggingface-hub==0.23.4
idna==3.6
imageio==2.34.1
imageio-ffmpeg==0.4.9
inflection==0.5.1
intel-openmp==2021.4.0
jax==0.4.28
jaxlib==0.4.28
Jinja2==3.1.4
jmespath==1.0.1
jsonpointer==2.4
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
mediapipe==0.10.14
mkl==2021.4.0
ml-dtypes==0.4.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
proglog==0.1.10
protobuf==4.25.3
psutil==5.9.8
psycopg2==2.9.9
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pybboxes==0.1.6
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rfc3987==1.3.8
roboflow==1.1.32
s3transfer==0.10.1
sahi==0.11.16
scipy==1.13.0
seaborn==0.13.2
shapely==2.0.4
six==1.16.0
smmap==5.0.1
sounddevice==0.4.6
sqlparse==0.4.4
strict-rfc3339==0.7
sympy==1.12.1
tbb==2021.12.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.32
ultralytics-thop==0.2.8
uritemplate==4.1.1
urllib3==2.2.1
validate_email==1.3
Werkzeug==3.0.3
